fixed small typo and update fltk readme

+++ Cpp/Stroustrup_principles_and_practice/lib/GUI/src/Graph.cpp
@@ -53,8 +53,8 @@ bool line_segment_intersect(Point p1, Point p2, Point p3, Point p4, Point& inter
    bool parallel;
    pair<double,double> u = line_intersect(p1,p2,p3,p4,parallel);
    if (parallel || u.first < 0 || u.first > 1 || u.second < 0 || u.second > 1) return false;
-   intersection.x = p1.x + u.first*(p2.x - p1.x);
-   intersection.y = p1.y + u.first*(p2.y - p1.y);
+   intersection.x = static_cast<int>(p1.x + u.first*(p2.x - p1.x));
+   intersection.y = static_cast<int>(p1.y + u.first*(p2.y - p1.y));
    return true;
 }
